requirements.txt
fastapi
uvicorn
pydantic
customtkinter
Pillow
pystray
PyJWT
python-dotenv
fastapi-sso
psutil
websockets

motor
passlib[bcrypt]
python-multipart
